+++ x/perp/v2/client/cli/cli_test.go
@@ -725,16 +725,28 @@ func (s *IntegrationTestSuite) TestDonateToEcosystemFund() {
 }
 
 func (s *IntegrationTestSuite) TestQueryModuleAccount() {
-	s.T().Logf("donate to ecosystem fund")
 	resp := new(types.QueryModuleAccountsResponse)
 	s.NoError(
-		testutilcli.ExecQuery(
-			s.network.Validators[0].ClientCtx,
-			cli.CmdQueryModuleAccounts(),
-			[]string{},
+		s.network.ExecQuery(
+			cli.NewQueryCmd(),
+			[]string{"module-accounts"},
 			resp,
 		),
 	)
+	s.NotEmpty(resp.Accounts)
+}
+
+func (s *IntegrationTestSuite) TestQueryCollateralDenom() {
+	resp := new(types.QueryCollateralResponse)
+	s.NoError(
+		s.network.ExecQuery(
+			cli.NewQueryCmd(),
+			[]string{"collateral"},
+			resp,
+		),
+	)
+	s.Equal(types.TestingCollateralDenomNUSD, resp.CollateralDenom,
+		"resp: %s", resp.String())
 }
 
 func TestIntegrationTestSuite(t *testing.T) {